options screen and updates
added an options screen and put the reset button there along with a message on how to alter tracking settings.
Added player HP to the map view and created a user default for it.
Included noble mask colours to the availible colours.
fixed how the game resets the user defaults so that it creates a list within a list for the masks.