considerable update - see details
Got multipul encounter spawns working at 25%.
Changed the movement area and timer rate to smaller and at a timer of 16 which seems to work better.
Made it so that there was a vibration when an encounter occours and this is turn-off-able in the options menu.
Made it so that you can wear the Avohkii or Vahi masks on your character sprite if you find them.