big annoying bug solving push
Solved the following bugs:
Fixed how far one has to move to spawn something
Fixed the freezing of animations occasionally when drawing a disk in the Rahi fight
Fixed another bug with how Rahi righting works to do with dodging after throwing.
Made it so that player masks dynamically updated when scrolled through on the player edit screen.